+++ CHANGELOG.md
@@ -41,11 +41,11 @@ _TBD_
 - [x] Added new properties to `AuthenticationToken`: `mode`, `invalidate`
 - [x] Added new properties to `Transaction`: `customFields`, `description`, `processedTime`, `scheduledTime`, `velocity`, `revision`
 - [x] Added new properties to `PaymentCard`: `method`, `panFingerprint`
-- [x] Added new properties to `BankAccount`: `accountNumberType`, `last4`, `bic`, `method`, `fingerprint`
+- [x] Added new properties to `BankAccount`: `accountNumberType`, `last4`, `bic`, `method`, `fingerprint`, `token`
 - [x] Added missing getters and setters to `AuthenticationToken`, `Transaction`, `PaymentCard`, `BankAccount`
 - [x] Removed `delete` method from `SessionService`
 - [x] Added new method to `SessionService`: `logout`
-- [x] Added new method to `BankAccountService`: `update`
+- [x] Added new method to `BankAccountService`: `update`, `createFromToken`
 
 ### Changed
 - [x] **Upgraded minimum PHP version to 7.1**